requirements.txt
PyQt6>=6.5
mss>=9.0
Pillow>=10.0
PyYAML>=6.0
numpy>=1.24
requests>=2.31
